+++ ee/hogai/graph/parallel_task_execution/mixins.py
@@ -49,7 +49,7 @@ class WithInsightCreationTaskExecution:
         The type allows None for compatibility with the base class.
         """
         # Import here to avoid circular dependency
-        from ee.hogai.graph.graph import InsightsAssistantGraph
+        from ee.hogai.graph.insights_graph.graph import InsightsGraph
 
         task = input_dict["task"]
         artifacts = input_dict["artifacts"]
@@ -57,7 +57,7 @@ class WithInsightCreationTaskExecution:
 
         self._current_task_id = task.id
 
-        # This is needed by the InsightsAssistantGraph to return an AssistantToolCallMessage
+        # This is needed by the InsightsGraph to return an AssistantToolCallMessage
         task_tool_call_id = f"task_{uuid.uuid4().hex[:8]}"
 
         formatted_instructions = AGENT_TASK_PROMPT_TEMPLATE.format(
@@ -73,7 +73,7 @@ class WithInsightCreationTaskExecution:
         )
 
         subgraph_result_messages: list[AssistantMessageUnion] = []
-        assistant_graph = InsightsAssistantGraph(self._team, self._user).compile_full_graph()
+        assistant_graph = InsightsGraph(self._team, self._user).compile_full_graph()
         try:
             async for chunk in assistant_graph.astream(
                 input_state,
@@ -148,7 +148,7 @@ class WithInsightCreationTaskExecution:
             if isinstance(message, VisualizationMessage) and message.id:
                 artifact = InsightArtifact(
                     task_id=task.id,
-                    id=None,  # The InsightsAssistantGraph does not create the insight objects
+                    id=None,  # The InsightsGraph does not create the insight objects
                     content=task.prompt,
                     query=cast(AnyAssistantGeneratedQuery, message.answer),
                 )
